Stop using 'project' to talk about an OSF 'node'
A project is no more than a node of a particular category. With child-nodes
and a more comprehensive use of OSF capabilities the discrepancy between
terminologies becomes more and more problematic. This change replaces
'project' with 'node' in all internal places, and the name of the
'project' parameter of the special remote.

This is also a sensible move as the created "projects" are actually
nodes of type "data" by default (not "project"), and the category is
configurable.