Remove crippledfs CI workflow
It fails permanently. However, neither does it fail on a real windows
box (crippled FS implied), nor in the windows CI workflow, nor on my
local system (with the same FAT image setup), nor on a fast exFAT
external SSD, nor on an ancient cheap USB stick, nor when putting my
local box under load in combination with all of the above
(`stress-ng --cpu 4 --io 2 --vm 1 --vm-bytes 1G --timeout 60s).

Whatever is happening in that workflow, it is not indicative of actual
behavior. There is no point in keeping it fail.